perf: optimize choco banana solver
